refactor(api/rpc): change permissions pkg to perms, rename NewReadOnlyClient to NewPublicClient, update tests for extra permissions level
